pyproject.toml
[tool.poetry]
name = "aicli"
version = "0.1.0"
description = ""
authors = ["Jordi Calafat <jcalafat97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.3.8"
python-dotenv = "^1.0.0"
rich = "^13.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
